requirements.txt
aiogram==3.1.1
lxml==4.9.3
openpyxl==3.1.2
pandas==2.1.1
python-dotenv==1.0.0
requests==2.31.0